So this is working
so maybe i just need the view that i created in the database, but i will keep the repos and all the models i have iam deleting the services if i need one i will create it

+++ Faza6/OnlineGamesShop/src/main/java/com/rich/onlinegamesshop/repository/CostumerRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface CostumerRepository extends JpaRepository<Costumer, Long>{
+public interface CostumerRepository extends JpaRepository<Costumer, Integer>{
 }
